pubsub-notify/message_sender/pyproject.toml
[tool.poetry]
name = "message-sender"
version = "0.1.0"
description = ""
authors = ["nownabe <nownabe@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-pubsub = "^2.13.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
